+++ module/testing/testing-everything-lad-lass.mjs
@@ -11,9 +11,6 @@ export function registerEverythingLadLass(quench) {
                 setQuenchTimeout(this);
 
                 describe("Everything Lad (5e)", function () {
-                    // On slower setups this can timeout. Increase the default timeout.
-                    this.timeout("20s");
-
                     const contents = `
                     <?xml version="1.0" encoding="UTF-16"?>
                     <CHARACTER version="6.0" TEMPLATE="builtIn.Superheroic.hdt">
@@ -968,9 +965,6 @@ export function registerEverythingLadLass(quench) {
                 });
 
                 describe("Everything Lass (6e)", function () {
-                    // On slower setups this can timeout. Increase the default timeout.
-                    this.timeout("20s");
-
                     const contents = `
                     <?xml version="1.0" encoding="UTF-16"?>
                       <CHARACTER version="6.0" TEMPLATE="builtIn.Superheroic6E.hdt">